g EM-353: Fixed by popping up a warning dialog if user attempts to delete a notification rule, which is used by a SSG cluster in monitoring property setup.
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@22804 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/ems/src/main/java/com/l7tech/server/ems/ui/pages/Monitor.java
@@ -105,7 +105,21 @@ public class Monitor extends EsmStandardWebPage {
             @Override
             protected Object getJsonResponseData() {
                 String guid = (String)deleteNotificationRuleDialog_id.getConvertedInput();
+                if (guid == null || guid.isEmpty()) {
+                    return new JSONException(new Exception("The selected notification rule is invalid."));
+                }
                 try {
+                    // Check if the notification rule is used by any cluster or nodes in monitoring property setups.
+                    for (SsgClusterNotificationSetup setup: ssgClusterNotificationSetupManager.findAll()) {
+                        String ssgClusterGuid = setup.getSsgClusterGuid();
+                        for (SystemMonitoringNotificationRule rule: setup.getSystemNotificationRules()) {
+                            if (guid.equals(rule.getGuid())) {
+                                String clusterName = ssgClusterManager.findByGuid(ssgClusterGuid).getName();
+                                return new JSONException(new Exception("Cannot delete '" + rule.getName() + "', since it is used by the cluster '" + clusterName + "' in monitoring properties setup."));
+                            }
+                        }
+                    }
+
                     String notificationRuleName = systemMonitoringNotificationRulesManager.findByGuid(guid).getName();
                     systemMonitoringNotificationRulesManager.deleteByGuid(guid);
